+++ src/main/java/com/ceyentra/visitor_management_system/rest/VisitorRestController.java
@@ -1,6 +1,7 @@
 package com.ceyentra.visitor_management_system.rest;
 
 import com.ceyentra.visitor_management_system.entity.Visitor;
+import com.ceyentra.visitor_management_system.exception.NotFoundException;
 import com.ceyentra.visitor_management_system.service.VisitorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -31,23 +32,41 @@ public class VisitorRestController {
 
     @PutMapping("/visitors/{visitorId}")
     public Visitor updateVisitor(@PathVariable int visitorId, @RequestBody Visitor visitor){
-        visitor.setVisitorId(visitorId);
-        return visitorService.updateVisitor(visitor);
+        if (visitorService.existsVisitor(visitorId)){
+            visitor.setVisitorId(visitorId);
+            return visitorService.updateVisitor(visitor);
+        }else {
+            throw new NotFoundException("Visitor not found with visitor ID : "+visitorId);
+        }
     }
 
     @DeleteMapping("/visitors/{visitorId}")
     public void deleteVisitor(@PathVariable int visitorId){
-        visitorService.deleteVisitor(visitorId);
+        if (visitorService.existsVisitor(visitorId)){
+            visitorService.deleteVisitor(visitorId);
+        }else {
+            throw new NotFoundException("Visitor not found with visitor ID : "+visitorId);
+        }
     }
 
     @GetMapping("/visitors/{visitorId}")
     public Visitor getVisitor(@PathVariable int visitorId){
-        return visitorService.findVisitorById(visitorId);
+        if (visitorService.existsVisitor(visitorId)){
+            return visitorService.findVisitorById(visitorId);
+        }else {
+            throw new NotFoundException("Visitor not found with visitor ID : "+visitorId);
+        }
+
     }
 
     @GetMapping("/visitors/nic/{nic}")
     public Visitor getVisitorByNic(@PathVariable String nic){
-        return visitorService.findVisitorByNic(nic);
+        Visitor visitorByNic = visitorService.findVisitorByNic(nic);
+        if (visitorByNic != null){
+            return visitorByNic;
+        }else {
+            throw new NotFoundException("Visitor not found with visitor NIC : "+nic);
+        }
     }
 
     @GetMapping("/visitors")
